Customer Update and Medical Info Added

+++ resources/views/admin/client/delegate/index.blade.php
@@ -63,7 +63,9 @@
                             @elseif($delegate->status == 0)
                               <a class="edit_option" href="#activeId{{ $delegate->id }}" data-toggle="modal"><i class="fas fa-caret-square-up"></i><span>Set Active</span></a>
                             @endif
+                            @if(Auth::check() && (Auth::user()->role == 'admin'))
                               <a class="delete_option" href="#delDelegate{{ $delegate->id }}" data-toggle="modal"><i class="fas fa-trash"></i><span>Delete</span></a>
+                            @endif
                             </td>
                 @include('admin.client.delegate.delegate_modal')
                 @include('admin.client.delegate.delegate_active')
