+++ resources/views/admin/location/district/index.blade.php
@@ -59,7 +59,9 @@
                             @elseif($district->status == 0)
                               <a class="edit_option" href="#activeId{{ $district->id }}" data-toggle="modal"><i class="fas fa-caret-square-up"></i><span>Set Active</span></a>
                             @endif
+                            @if(Auth::check() && (Auth::user()->role == 'admin'))
                               <a class="delete_option" href="#delDistrict{{ $district->id }}" data-toggle="modal"><i class="fas fa-trash"></i><span>Delete District</span></a>
+                            @endif
                             </td>
                 @include('admin.location.district.district_modal')
                 @include('admin.location.district.district_active')
