+++ resources/views/admin/visa/link/index.blade.php
@@ -59,7 +59,9 @@
                             @elseif($link->status == 0)
                               <a class="edit_option" href="#activeId{{ $link->id }}" data-toggle="modal"><i class="fas fa-caret-square-up"></i><span>Set Active</span></a>
                             @endif
+                            @if(Auth::check() && (Auth::user()->role == 'admin'))
                               <a class="delete_option" href="#delLink{{ $link->id }}" data-toggle="modal"><i class="fas fa-trash"></i><span>Delete</span></a>
+                            @endif
                             </td>
                 @include('admin.visa.link.link_modal')
                 @include('admin.visa.link.link_active')
